Apply random forest
A random forest model was fit to the data, with m = p/3 = 10. 
The MSE for this model was 15.03455, which was lower than for both earlier models, indicating that random forests yielded an improvement.
The varImpPlot( ) and the importance( ) functions indicated that, again, ‘failures’ and ‘absences’ were by far the two most important variables in determining the final grade (G3) of a student’s math course.